+++ wazo_call_logd/database/queries/__init__.py
@@ -1,0 +1,15 @@
+# Copyright 2020 The Wazo Authors  (see the AUTHORS file)
+# SPDX-License-Identifier: GPL-3.0-or-later
+
+from .call_log import CallLogDAO
+
+
+class DAO:
+
+    _daos = {
+        'call_log': CallLogDAO,
+    }
+
+    def __init__(self, session):
+        for name, dao in self._daos.items():
+            setattr(self, name, dao(session))
