+++ wazo_call_logd/plugins/cdr/plugin.py
@@ -3,8 +3,8 @@
 
 from wazo_auth_client import Client as AuthClient
 
-from wazo_call_logd.database.dao import new_db_session
-from wazo_call_logd.database.dao import CallLogDAO
+from wazo_call_logd.database.helpers import new_db_session
+from wazo_call_logd.database.queries import DAO
 
 from .resource import CDRResource
 from .resource import CDRIdResource
@@ -19,7 +19,7 @@ class Plugin:
         config = dependencies['config']
 
         auth_client = AuthClient(**config['auth'])
-        dao = CallLogDAO(new_db_session(config['db_uri']))
+        dao = DAO(new_db_session(config['db_uri'])).call_log
         service = CDRService(dao)
 
         api.add_resource(CDRResource, '/cdr', resource_class_args=[service])
